Expose/test WithCancellation for async enumerables
